% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_messages.R
\name{abort_bad_argument}
\alias{abort_bad_argument}
\title{Create an error message for bad arguments}
\usage{
abort_bad_argument(arg, must, not = NULL, call = rlang::caller_env())
}
\arguments{
\item{arg}{Character Argument of function error relates to}

\item{must}{What must the function be.}

\item{not}{Argument that error is about. Optional, by default NULL}

\item{call}{Returns the parent environment that called the error with
\code{rlang::caller_env()}.}
}
\value{
None. Throws error_bad_argument error
}
\description{
Create an error message for bad arguments
}
\details{
Taken from Wickham, H. (2019). Advanced R, Second Edition (2nd ed.).
Chapman and Hall/CRC. https://doi.org/10.1201/9781351201315
}
