% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_messages.R
\name{abort_missing_selection}
\alias{abort_missing_selection}
\title{Create an error message for missing selection}
\usage{
abort_missing_selection(arg, not_in, call = rlang::caller_env())
}
\arguments{
\item{arg}{Character Argument of function error relates to.}

\item{not_in}{Object that arg is missing from.}

\item{call}{Returns the parent environment that called the error with
\code{rlang::caller_env()}.}
}
\value{
None. Throws error_missing_selection error
}
\description{
Create an error message for missing selection
}
