% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_data_variable_present}
\alias{check_data_variable_present}
\title{Check if a literal variable is present in the data context.}
\usage{
check_data_variable_present(var, data, call_level = 1)
}
\arguments{
\item{var}{Literal variable to check}

\item{data}{Data context of variable}

\item{call_level}{level of parent environment to show as the error call level}
}
\description{
Check if a literal variable is present in the data context.
}
