% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElectionR-package.R
\docType{package}
\name{ElectionR-package}
\alias{ElectionR}
\alias{ElectionR-package}
\title{ElectionR: Analyse Election Results}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Ben Hamiton \email{bjh685@gmail.com}

}
\keyword{internal}
