% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote_share.R
\name{vote_share}
\alias{vote_share}
\title{Calculate the vote share of parties}
\usage{
vote_share(data, .party = NULL, .raw_vote = NULL)
}
\arguments{
\item{data}{A data frame of parties and their raw vote numbers}

\item{.party}{Character vector of column name that contains party names. By
default the first column}

\item{.raw_vote}{Character vector of column name that contain raw vote numbers.
By default the second column}
}
\value{
An object of the same type as parties. The output has the following properties:
\itemize{
\item A new column: vote_share that contains the vote share of each party.
}
}
\description{
Calculate the vote share of parties
}
\examples{
df <- data.frame(party = c("A Group", "B Team"), votes = c(5, 9))
vote_share(df, "party", "votes")

}
